refact(tests): move upgrade tests into `upgrade_package` dir
New venv management tests will go into a separate testing directory
