+++ upgrade/tests/upgrade_package/test_check_constraints_file_path.py
@@ -2,7 +2,7 @@ import pytest
 from pathlib import Path
 from mock import patch
 from upgrade.scripts.upgrade_python_package import get_constraints_file_path
-from .conftest import VENV_PATH, install_local_package
+from upgrade.tests.upgrade_package.conftest import VENV_PATH, install_local_package
 
 
 @pytest.mark.parametrize(
